chore(ln-dlc-node): Rename multiple_payments test module
Calling it `payments.rs` is better since we want to add more tests
related to payments through a JIT channel.
