+++ crates/ln-dlc-node/src/tests/just_in_time_channel/mod.rs
@@ -1,4 +1,4 @@
 mod channel_close;
 mod create;
 mod fail_intercepted_htlc;
-mod multiple_payments;
+mod payments;
